Stuck on port number
Br. Burton, in class today I used this file to set up my database.

I got stuck for a large portion of the class. The reason was due to the port number (apparently '5433' was the default port number for my computer, but the code specified '5432'). I'm proposing a pull request because this file is a year old and this might be a fairly recent change.

I am on Windows and using postgres 9.6.